+++ CHANGELOG.md
@@ -14,6 +14,9 @@ directly impact users rather than highlighting other crucial architectural updat
 ### Removed
 - The sapling address and the QR of it has been removed from the receive screen. The only meaningful options are the UA and the transparent addresses. 
 
+### Added
+- Confirmation screen when sending funds. The initial screen is about filling in the address, amount and message (optional). The butoon `review` leads to a brand new screen where the summary of the transaction is presented. The send is confirmed by tapping the send button. Going back to update send data is possible via `go back` button.
+
 ## 0.2.0 build 9 (2023-11-14)
 
 ### Changed